git - update path to gpg
Previous path was pointing to a MacOS specific tool. Updated it with the generic gpg tool.
Note that in my Arch Linux install, gpg is located in: `/usr/bin/gpg` but on MacOS, it's in `/usr/local/bin/gpg`.
You can create a symlink to resolve this issue (On Arch):
```sh
ln -s /usr/bin/gpg /usr/local/bin/gpg
```