fish - create/attach to tmux by default & misc
Automatically create or attach to a tmux session when shell is starting.
We first check if `TMUX` variable is set to avoid nesting.
Then we create a new session with the `-A` flag, it will either create a new session or attach to an existing one.
This need to be in a separate block otherwise the abbreviations and the rest will not be set inside tmux.

Misc: move alias and environment variable inside `is-interactive` shell.

[autostart tmux](https://www.markhansen.co.nz/auto-start-tmux/)
[tmux create or attach](https://unix.stackexchange.com/questions/103898/how-to-start-tmux-with-attach-if-a-session-exists/176885#176885)